Move usePaginatedList logic
