+++ components/pages/organizations/list/entity-list.tsx
@@ -10,7 +10,7 @@ import { getVoteStatus } from '@lib/util'
 
 import { ELECTIONS_DETAILS } from '@const/routes'
 import RouterService from '@lib/router'
-import { PaginatedListTemplate } from '@components/pages/app/page-templates/paginated-list-template'
+import { PaginatedListTemplate, usePaginatedList } from '@components/pages/app/page-templates/paginated-list-template'
 import { EntitiesFilter, IFilterEntity } from '../components/entities-filter'
 import { useEntityList } from '@hooks/use-entities'
 
@@ -20,16 +20,10 @@ interface IDashboardProcessListProps {
   totalCount?: number
 }
 
-/** This sets pagination next offset for process pagination */
-const PAGINATION_FROM = 64
-
 export const DashboardEntityList = ({
   pageSize = 8,
   totalCount = 0,
 }: IDashboardProcessListProps) => {
-  ///////////////////////////////
-  // JSX
-  ///////////////////////////////
 
   // Render item on the list from it summary
   const renderProcessItem = (identity: string) => {
@@ -79,8 +73,8 @@ export const DashboardEntityList = ({
     filter: filter,
     setFilter: setFilter,
     dataList: entitiesList,
-    dataPagination,
-    setDataPagination,
+    backendDataPagination: dataPagination,
+    setBackendDataPagination: setDataPagination,
   })
 
   return (
@@ -111,88 +105,3 @@ export const DashboardEntityList = ({
   )
 }
 
-// todo: move this arguments to generic types
-// export const usePaginatedList = ({
-
-interface IUsePaginatedListProps<Filter, DataList> {
-  filter: Filter
-  setFilter: (Filter) => void
-  dataList: DataList
-  dataPagination: number
-  setDataPagination: (number) => void
-}
-
-export function usePaginatedList <F, DataList>({
-  filter,
-  setFilter,
-  dataList,
-  dataPagination,
-  setDataPagination,
-}: IUsePaginatedListProps<F, DataList>) {
-
-  const [cachedData, setCachedData] = useState<DataList[]>([])
-
-  ///////////////////////////////
-  // PAGINATOR
-  ///////////////////////////////
-
-  // Set the page at initial state
-  const resetPage = useCallback(() => {
-    setCurrentPage(1)
-    setCachedData([])
-  }, [])
-
-  // Paginator current page
-  const [currentPage, setCurrentPage] = useState(1)
-
-  // When processIds are retrieved, update the list of already loaded process ids
-  // Used for pagination, if we need to load next 64 processes
-  useEffect(() => {
-    // if (loading != true) setLoading(true)
-    setCachedData(cachedData.concat(dataList))
-  }, [dataList])
-
-  const [renderedData, setRenderedData] = useState<DataList[]>([])
-
-  const loadMoreData = () => {
-    setDataPagination(dataPagination + PAGINATION_FROM)
-  }
-
-  ///////////////////////////////
-  // Filter
-  ///////////////////////////////
-
-  const filterIsChanged = (filter, tempFilter) =>
-    JSON.stringify(filter) !== JSON.stringify(tempFilter)
-
-  const enableFilter = (tempFilter) => {
-    if (filterIsChanged(filter, tempFilter)) {
-      resetPage()
-      setFilter(Object.assign({}, tempFilter))
-    }
-  }
-
-  const disableFilter = (tempFilter, resetForm: { (): void }) => {
-    resetForm()
-    if (
-      Object.keys(filter).length !== 0 // Check if filter is already reset
-    ) {
-      resetPage()
-      setFilter({})
-    }
-  }
-
-  return {
-    cachedData,
-    renderedData,
-    currentPage,
-    dataPagination,
-    methods: {
-      enableFilter,
-      disableFilter,
-      setRenderedData,
-      setCurrentPage,
-      loadMoreData,
-    },
-  }
-}
